samples/python-markdown/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.4
azure-identity==1.14.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Deprecated==1.2.14
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.4
microsoft-kiota-abstractions==0.8.6
microsoft-kiota-authentication-azure==0.3.0
microsoft-kiota-http==0.6.1
microsoft-kiota-serialization-json==0.4.1
microsoft-kiota-serialization-text==0.2.1
msal==1.24.1
msal-extensions==1.0.0
msgraph-core==1.0.0a4
msgraph-sdk==1.0.0a16
multidict==6.0.4
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-frontmatter==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
std-uritemplate==0.0.42
typing_extensions==4.8.0
urllib3==2.0.6
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
